Rename `exerciseSingle` to `exerciseDocument`
It makes more sense to use the term 'document' to describe the purpose of this module, since it's consistent with the terminology used by both `vuexfire` but also `firestore` itself.

On the other hand, 'single' was a term used only by us here in these Vuex modules, and it seemed less meaningful in context.
